fix: unify My Trades chat selection
Use one shared projection for the My Trades sidebar and Enter/send handlers to prevent selection/action desync.
